ai-reasoning-service/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
openai==1.12.0
langchain==0.1.9
langchain-openai==0.0.8
langchain-community==0.0.24
langgraph==0.0.26
pydantic==2.6.1
python-dotenv==1.0.1
httpx==0.26.0
pgvector==0.2.4
psycopg2-binary==2.9.9
sqlalchemy==2.0.27
numpy==1.26.4